feat: add insert/update methods in FilmDbStorage

+++ src/main/resources/schema.sql
@@ -7,13 +7,13 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS requested_friends (
-    requested_friends_id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    requested_friends_id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     user_id INT REFERENCES users(user_id),
     friend_id INT REFERENCES users(user_id)
 );
 
 CREATE TABLE IF NOT EXISTS approved_friends (
-    approved_friends_id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    approved_friends_id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     user1_id INT REFERENCES users(user_id),
     user2_id INT REFERENCES users(user_id)
 );
@@ -34,17 +34,17 @@ CREATE TABLE IF NOT EXISTS films (
     description VARCHAR(200),
     release_date DATE,
     duration INT,
-    rating_id BIGINT REFERENCES rating(rating_id)
+    rating_id INT REFERENCES rating(rating_id)
 );
 
 CREATE TABLE IF NOT EXISTS films_genres (
-    films_genres BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    films_genres INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     film_id INT REFERENCES films(film_id),
     genre_id INT REFERENCES genres(genres_id)
 );
 
 CREATE TABLE IF NOT EXISTS films_likes (
-    films_likes BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    films_likes INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     film_id INT REFERENCES films(film_id),
     user_id INT REFERENCES users(user_id)
 );
